+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.15.3] - (18.06.2024)
+
+## Changes
+
+- polling code refactor, added impulse generator
+- remove production properties from Q-Relay accessory
+- cleanup
+
 ## [7.15.2] - (09.06.2024)
 
 ## Changes